Rename package utils
