+++ src/Code/Screen/FrictionCoeficient.java
@@ -1,6 +1,6 @@
 package Code.Screen;
 import Code.Object.Objectss;
-import Code.LinhTinh.ExceptionCase;
+import Code.Utils.ExceptionCase;
 
 import javax.swing.*;
 import javax.swing.border.Border;
